Refactored API to allow easier versioning
